src/utils/config_loader.py
# src/utils/config_loader.py
import yaml
from pathlib import Path

PROJECT_ROOT = Path(__file__).resolve().parents[2]

def load_config(path: str = None) -> dict:
    cfg_path = path or PROJECT_ROOT / "config" / "config.yaml"
    with open(cfg_path) as f:
        return yaml.safe_load(f)
