requirements.txt
tensorflow==2.15.0
keras==2.15.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
matplotlib==3.8.3
seaborn==0.13.2
requests==2.31.0
beautifulsoup4==4.12.3
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.0.1
gradio==4.26.0
shap==0.44.0
mlflow==2.11.0
pytest==8.0.0
lxml==5.1.0
urllib3==2.2.1
scipy==1.13.0
pyarrow==15.0.0
